bussiness1/InventoryManagement-main/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
WTForms==3.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Jinja2>=3.1.2 